chore(web): extract `form.util` and `headers` from `index.ts`
