+++ packages/web/src/common/utils/index.ts
@@ -1,30 +1,7 @@
 import {
   ID_CONTEXT_MENU_ITEMS,
-  ID_EVENT_FORM,
   ID_SOMEDAY_EVENT_FORM,
 } from "../constants/web.constants";
-
-export const headers = (token?: string) => {
-  if (token) {
-    return {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    };
-  } else {
-    return {
-      headers: {
-        Authorization: `Bearer ${localStorage.getItem(
-          "removedAsPartOfTokenRefactor",
-        )}`,
-      },
-    };
-  }
-};
-
-export const isEventFormOpen = () =>
-  document.getElementsByName(ID_EVENT_FORM).length === 1 ||
-  document.getElementsByName(ID_SOMEDAY_EVENT_FORM).length === 1;
 
 export const isSomedayEventFormOpen = () =>
   document.getElementsByName(ID_SOMEDAY_EVENT_FORM).length === 1;
