+++ packages/web/src/views/Calendar/hooks/shortcuts/useShortcuts.ts
@@ -10,9 +10,9 @@ import {
 import { Categories_Event } from "@core/types/event.types";
 import { ROOT_ROUTES } from "@web/common/constants/routes";
 import { ID_REMINDER_INPUT } from "@web/common/constants/web.constants";
-import { isEventFormOpen } from "@web/common/utils";
 import { createAlldayDraft } from "@web/common/utils/draft/draft.util";
 import { createTimedDraft } from "@web/common/utils/draft/draft.util";
+import { isEventFormOpen } from "@web/common/utils/form/form.util";
 import {
   selectIsAtMonthlyLimit,
   selectIsAtWeeklyLimit,
